requirements.txt
appdirs==1.4.3
asgiref==3.2.3
asn1crypto==0.22.0
awscli==1.11.109
awsebcli==3.10.3
backports.ssl==0.0.9
blessed==1.14.2
boto==2.47.0
boto3==1.4.4
botocore==1.5.77
cement==2.8.2
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
colorama==0.3.7
defusedxml==0.6.0
gunicorn==20.0.4
Django==3.0.2
django-admin-honeypot==1.1.0
django-allauth==0.41.0
django-bootstrap3==12.0.3
django-jquery==3.1.0
django-jquery-ui==1.11.4.1
django-ses==0.8.14
django-storages==1.8
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
future==0.18.2
idna==2.5
jmespath==0.9.3
oauthlib==3.1.0
pathspec==0.5.0
Pillow==7.0.0
psycopg2-binary==2.8.4
pyasn1==0.2.3
pycparser==2.17
python-dateutil==2.6.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==3.12
requests==2.22.0
requests-oauthlib==1.3.0
rsa==3.4.2
s3transfer==0.1.10
semantic-version==2.5.0
six==1.10.0
sqlparse==0.3.0
stripe==2.42.0
tabulate==0.7.5
termcolor==1.1.0
urllib3==1.25.8
wcwidth==0.1.7
websocket-client==0.44.0
